extract base serve code
