+++ lib/src/main/java/com/diffplug/spotless/npm/PrettierFormatterStep.java
@@ -70,7 +70,9 @@ public class PrettierFormatterStep {
 									NpmResourceHelper.readUtf8StringFromClasspath(PrettierFormatterStep.class, "/com/diffplug/spotless/npm/prettier-package.json"),
 									new TreeMap<>(devDependencies)),
 							"prettier",
-							NpmResourceHelper.readUtf8StringFromClasspath(PrettierFormatterStep.class, "/com/diffplug/spotless/npm/prettier-serve.js"),
+							NpmResourceHelper.readUtf8StringFromClasspath(PrettierFormatterStep.class,
+									"/com/diffplug/spotless/npm/common-serve.js",
+									"/com/diffplug/spotless/npm/prettier-serve.js"),
 							npmPathResolver.resolveNpmrcContent()),
 					buildDir,
 					npmPathResolver.resolveNpmExecutable());
